+++ backend/src/controller/user.controller.ts
@@ -1,7 +1,7 @@
 import { RowDataPacket } from "mysql2";
 import randomString from "randomstring";
 
-import { TABLE_NAME, CreateUserModel, LoginModel, UserColumn, createUserSql } from "../model/user.model";
+import { USER_TABLE_NAME, CreateUserModel, UserColumn, createUserSql } from "../model/user.model";
 import { insert, select, OptionType } from "../util/sql";
 import { createDigest, checkPassword } from "../util/password";
 import UnauthorizedError from "../error/unauthorized";
@@ -20,7 +20,7 @@ const controller = {
             });
 
             const options: OptionType = {
-                table: TABLE_NAME,
+                table: USER_TABLE_NAME,
                 column: [UserColumn.code],
                 limit: 1,
                 where: `${UserColumn.code} = "${code}"`
@@ -40,28 +40,6 @@ const controller = {
         const response = await insert(sql);
 
         return response;
-    },
-    login: async (data: JSON) => {
-        const model: LoginModel = Object.assign(data);
-        const options: OptionType = {
-            table: TABLE_NAME,
-            column: [UserColumn.password],
-            limit: 1,
-            where: `${UserColumn.email} = "${model.email}"`
-        };
-
-        const response: RowDataPacket[] = await select(options);
-
-        if (response.length <= 0) {
-            throw new UnauthorizedError("Invalid Email");
-        }
-
-        const password: string = response[0].password;
-        const isCheck: boolean = await checkPassword(model.password, password);
-
-        if (!isCheck) throw new UnauthorizedError("Invalid Password");
-
-        console.log(isCheck);
     }
 };
