+++ backend/src/model/auth.model.ts
@@ -1,0 +1,11 @@
+export interface LoginModel {
+    email: string;
+    password: string;
+}
+
+export interface LoginResponseModel {
+    accessToken: string;
+    accessTokenExpiredAt: number; // unix timestamp
+    refreshToken: string;
+    refreshTokenExpiredAt: number; // unix timestamp
+}
